+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [0.12.0] - 2022-01-16
 
 ### added
 
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### changed
 
-- [x] `Pfx.address/1` now properly returns a full address  prefix, always without mask
+- `Pfx.address/1` now properly returns a full address  prefix, always without mask
     - example: address({{1, 1, 1, 1}, 24}) -> {1, 1, 1, 1}